# .git-rewrite/t/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PetPlantr API",
    "description": "AI-powered dog photo to 3D planter conversion",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/breed/detect": {
      "post": {
        "tags": [
          "breed-detection",
          "breed-detection"
        ],
        "summary": "Detect Breed",
        "description": "Detect dog breed from image with high accuracy\nSupports both URL and base64 image inputs",
        "operationId": "detect_breed_api_v1_breed_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreedDetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreedDetectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/breed/detect-file": {
      "post": {
        "tags": [
          "breed-detection",
          "breed-detection"
        ],
        "summary": "Detect Breed From File",
        "description": "Detect breed from uploaded file",
        "operationId": "detect_breed_from_file_api_v1_breed_detect_file_post",
        "parameters": [
          {
            "name": "use_tta",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Use Tta"
            }
          },
          {
            "name": "confidence_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.8,
              "title": "Confidence Threshold"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_detect_breed_from_file_api_v1_breed_detect_file_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/breed/batch-detect": {
      "post": {
        "tags": [
          "breed-detection",
          "breed-detection"
        ],
        "summary": "Batch Detect Breeds",
        "description": "Batch breed detection for multiple images",
        "operationId": "batch_detect_breeds_api_v1_breed_batch_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/BreedDetectionRequest"
                },
                "type": "array",
                "title": "Requests"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/breed/calibrate": {
      "post": {
        "tags": [
          "breed-detection",
          "breed-detection"
        ],
        "summary": "Calibrate Confidence",
        "description": "Recalibrate confidence scores based on validation data",
        "operationId": "calibrate_confidence_api_v1_breed_calibrate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "object"
                },
                "type": "array",
                "title": "Calibration Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_detect_breed_from_file_api_v1_breed_detect_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_detect_breed_from_file_api_v1_breed_detect_file_post"
      },
...
      "BreedDetectionRequest": {
        "properties": {
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64"
          },
          "use_tta": {
            "type": "boolean",
            "title": "Use Tta",
            "default": true
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold",
            "default": 0.8
          },
          "return_top_k": {
            "type": "integer",
            "title": "Return Top K",
            "default": 5
          }
        },
        "type": "object",
        "title": "BreedDetectionRequest",
        "description": "Request model for breed detection"
      },
      "BreedDetectionResponse": {
        "properties": {
          "predicted_breed": {
            "type": "string",
            "title": "Predicted Breed"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "processing_time": {
            "type": "number",
            "title": "Processing Time"
          },
          "top_predictions": {
            "items": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Top Predictions"
          },
          "is_high_confidence": {
            "type": "boolean",
            "title": "Is High Confidence"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "predicted_breed",
          "confidence",
          "processing_time",
          "top_predictions",
          "is_high_confidence",
          "model_version",
          "metadata"
        ],
        "title": "BreedDetectionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Problem": {
        "type": "object",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "description": "RFC 7807 Problem Details"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}